App: Add profiles store to maintain all seen user profiles
