+++ app/src/pages/users/_id.vue
@@ -9,20 +9,12 @@ import { mapState } from 'vuex';
 import MainWrapper from '../../components/MainWrapper.vue';
 import with404Handler from '~/util/with404Handler';
 
-const isCurrentUser = (store, route) =>
-  store.state.currentUser?.name === route.params.id;
-
 export default {
   components: { MainWrapper },
 
   asyncData({ store, route, error }) {
-    if (isCurrentUser(store, route)) {
-      // if we're going to the current user's page, we don't need to load color
-      // scheme. by avoiding this extra load we prevent a visual bug when
-      // navigating to current user profile on mobile nav bar.
-      //
-      // this all goes away once color schemes are present before navigation
-      // happens.
+    if (store.state.profiles[route.params.id]) {
+      // if we've already loaded this profile, skip this
       return;
     }
 
@@ -35,11 +27,7 @@ export default {
   computed: {
     ...mapState({
       colorScheme(state) {
-        if (isCurrentUser(this.$store, this.$route)) {
-          return state.currentUser.colorScheme;
-        } else {
-          return state.profile.user.colorScheme;
-        }
+        return state.profiles[this.$route.params.id]?.colorScheme;
       },
     }),
   },
